+++ app/src/main/kotlin/com/getstrm/pace/processing_platforms/ProcessingPlatformViewGenerator.kt
@@ -12,6 +12,7 @@ import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.FieldT
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.FieldTransform.Transform.TransformCase.TRANSFORM_NOT_SET
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.Filter.FilterCase.GENERIC_FILTER
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.Filter.FilterCase.RETENTION_FILTER
+import com.getstrm.pace.processing_platforms.DefaultProcessingPlatformTransformer.renderName
 import com.getstrm.pace.util.defaultJooqSettings
 import com.getstrm.pace.util.fullName
 import com.getstrm.pace.util.headTailFold
@@ -52,7 +53,11 @@ abstract class ProcessingPlatformViewGenerator(
             field(
                 "{0} > {1}",
                 Boolean::class.java,
-                timestampAdd(field(unquotedName(field.fullName()), Timestamp::class.java), this.period.days, DatePart.DAY),
+                timestampAdd(
+                    field(unquotedName(field.fullName()), Timestamp::class.java),
+                    this.period.days,
+                    DatePart.DAY
+                ),
                 currentTimestamp()
             )
         } else {
@@ -61,15 +66,17 @@ abstract class ProcessingPlatformViewGenerator(
 
     protected open val jooq: DSLContext = using(SQLDialect.DEFAULT, defaultJooqSettings.apply(customJooqSettings))
 
-    fun toDynamicViewSQL(): String {
+    open fun createOrReplaceView(name: String) = jooq.createOrReplaceView(name)
+
+    open fun toDynamicViewSQL(): Queries {
         val queries = dataPolicy.ruleSetsList.map { ruleSet ->
             val targetView = ruleSet.target.fullname
-            jooq.createOrReplaceView(renderName(targetView)).`as`(toSelectStatement(ruleSet))
+            createOrReplaceView(renderName(targetView)).`as`(toSelectStatement(ruleSet))
         }
 
         val allQueries = queries + additionalFooterStatements()
 
-        return jooq.queries(allQueries).sql
+        return jooq.queries(allQueries)
     }
 
     fun toSelectStatement(ruleSet: DataPolicy.RuleSet) =
@@ -87,14 +94,17 @@ abstract class ProcessingPlatformViewGenerator(
                 addDetokenizeJoins(it, ruleSet)
             }
             .where(
-                ruleSet.filtersList.map { filter ->
-                    when (filter.filterCase) {
-                        RETENTION_FILTER -> toCondition(filter.retentionFilter)
-                        GENERIC_FILTER -> toCondition(filter.genericFilter)
-                        else -> throw IllegalArgumentException("Unsupported filter: ${filter.filterCase.name}")
-                    }
-                }
+                createWhereStatement(ruleSet)
             )
+
+    private fun createWhereStatement(ruleSet: DataPolicy.RuleSet) =
+        ruleSet.filtersList.map { filter ->
+            when (filter.filterCase) {
+                RETENTION_FILTER -> toCondition(filter.retentionFilter)
+                GENERIC_FILTER -> toCondition(filter.genericFilter)
+                else -> throw IllegalArgumentException("Unsupported filter: ${filter.filterCase.name}")
+            }
+        }
 
     private fun addDetokenizeJoins(
         selectJoinStep: SelectJoinStep<Record>,
@@ -118,7 +128,7 @@ abstract class ProcessingPlatformViewGenerator(
         return result
     }
 
-    fun toCondition(filter: DataPolicy.RuleSet.Filter.GenericFilter): Condition {
+    open fun toCondition(filter: DataPolicy.RuleSet.Filter.GenericFilter): Condition {
         if (filter.conditionsList.size == 1) {
             // If there is only one filter it should be the only option
             return getParser().parseCondition(filter.conditionsList.first().condition)
@@ -147,7 +157,7 @@ abstract class ProcessingPlatformViewGenerator(
         return DSL.condition(whereCondition)
     }
 
-    fun toCondition(retention: DataPolicy.RuleSet.Filter.RetentionFilter): Condition {
+    open fun toCondition(retention: DataPolicy.RuleSet.Filter.RetentionFilter): Condition {
         if (retention.conditionsList.size == 1) {
             // If there is only one filter it should be the only option
             // create retention sql