+++ app/src/main/kotlin/com/getstrm/pace/processing_platforms/bigquery/BigQueryClient.kt
@@ -59,7 +59,7 @@ class BigQueryClient(
 
     override suspend fun applyPolicy(dataPolicy: DataPolicy) {
         val viewGenerator = BigQueryViewGenerator(dataPolicy, userGroupsTable)
-        val query = viewGenerator.toDynamicViewSQL()
+        val query = viewGenerator.toDynamicViewSQL().sql
         val queryConfig = QueryJobConfiguration.newBuilder(query)
             .setUseLegacySql(false)
             .build()