+++ app/src/main/kotlin/com/getstrm/pace/processing_platforms/synapse/SynapseClient.kt
@@ -1,30 +1,31 @@
-package com.getstrm.pace.processing_platforms.postgres
+package com.getstrm.pace.processing_platforms.synapse
 
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy
-import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.ProcessingPlatform.PlatformType.POSTGRES
-import com.getstrm.pace.config.PostgresConfig
+import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.ProcessingPlatform.PlatformType.SYNAPSE
+import com.getstrm.pace.config.SynapseConfig
 import com.getstrm.pace.processing_platforms.Group
 import com.getstrm.pace.processing_platforms.ProcessingPlatformClient
 import com.getstrm.pace.processing_platforms.Table
+import com.getstrm.pace.processing_platforms.databricks.SynapseViewGenerator
 import com.getstrm.pace.util.normalizeType
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.jooq.DSLContext
+import org.jooq.Field
 import org.jooq.SQLDialect
 import org.jooq.impl.DSL
-import org.slf4j.LoggerFactory
 
-class PostgresClient(
+class SynapseClient(
     override val id: String,
     private val jooq: DSLContext,
 ) : ProcessingPlatformClient {
     // To match the behavior of the other ProcessingPlatform implementations, we connect to a
-    // single database. If we want to add support for a single client to connect to mulitple
+    // single database. If we want to add support for a single client to connect to multiple
     // databases, more info can be found here:
     // https://www.codejava.net/java-se/jdbc/how-to-list-names-of-all-databases-in-java
-    constructor(config: PostgresConfig) : this(
+    constructor(config: SynapseConfig) : this(
         config.id,
         DSL.using(
             HikariDataSource(
@@ -33,35 +34,35 @@ class PostgresClient(
                     username = config.userName
                     password = config.password
                 }),
-            SQLDialect.POSTGRES
-        )
+            SQLDialect.DEFAULT
+        ),
     )
 
-    private val log by lazy { LoggerFactory.getLogger(javaClass) }
-
-    override val type = POSTGRES
+    override val type = SYNAPSE
 
     override suspend fun listTables(): List<Table> = jooq.meta()
         .filterSchemas { !schemasToIgnore.contains(it.name) }
         .tables
-        .map { PostgresTable(it) }
+        .map { SynapseTable(it) }
 
     override suspend fun applyPolicy(dataPolicy: DataPolicy) {
-        val viewGenerator = PostgresViewGenerator(dataPolicy)
-        val query = viewGenerator.toDynamicViewSQL().sql
+        val viewGenerator = SynapseViewGenerator(dataPolicy)
+        // Explicitly dropping the views if they exist first and granting the 'select' to roles is required by Synapse.
+        val dropQuery = viewGenerator.dropViewsSQL().queries()
+        val query = viewGenerator.toDynamicViewSQL().queries()
+        val grantSelectQuery = viewGenerator.grantSelectPrivileges().queries()
 
         withContext(Dispatchers.IO) {
-            jooq.query(query).execute()
+            jooq.queries(*dropQuery, *query, *grantSelectQuery).executeBatch()
         }
     }
 
     override suspend fun listGroups(): List<Group> {
         val result = withContext(Dispatchers.IO) {
-            jooq.select(DSL.field("oid", Int::class.java), DSL.field("rolname", String::class.java))
-                .from(DSL.table("pg_roles"))
+            jooq.select(DSL.field("principal_id", Int::class.java), DSL.field("name", String::class.java))
+                .from(DSL.table("sys.database_principals"))
                 .where(
-                    DSL.field("rolcanlogin").notEqual(true),
-                    DSL.field("rolname").notLike("pg_%")
+                    DSL.field("type_desc").eq("DATABASE_ROLE")
                 )
                 .fetch()
         }
@@ -70,12 +71,12 @@ class PostgresClient(
     }
 
     companion object {
-        // These are built-in Postgres schemas that we don't want to list tables from.
-        private val schemasToIgnore = listOf("information_schema", "pg_catalog")
+        // These are built-in Synapse schemas that we don't want to list tables from.
+        private val schemasToIgnore = listOf("sys", "INFORMATION_SCHEMA")
     }
 }
 
-class PostgresTable(
+class SynapseTable(
     val table: org.jooq.Table<*>
 ) : Table() {
     override val fullName: String = "${table.schema?.name}.${table.name}"