+++ app/src/test/kotlin/com/getstrm/pace/processing_platforms/synapse/SynapseViewGeneratorTest.kt
@@ -1,13 +1,20 @@
-package com.getstrm.pace.processing_platforms.databricks
+package com.getstrm.pace.processing_platforms.synapse
 
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy
+import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.Principal
+import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.FieldTransform
 import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.Filter.GenericFilter
+import build.buf.gen.getstrm.pace.api.entities.v1alpha.DataPolicy.RuleSet.Filter.RetentionFilter
+import com.getstrm.pace.exceptions.BadRequestException
+import com.getstrm.pace.exceptions.InternalException
 import com.getstrm.pace.namedField
+import com.getstrm.pace.processing_platforms.databricks.SynapseViewGenerator
 import com.getstrm.pace.toPrincipal
 import com.getstrm.pace.toPrincipals
 import com.getstrm.pace.toSql
 import com.getstrm.pace.util.parseDataPolicy
 import com.getstrm.pace.util.yaml2json
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.shouldBe
 import org.intellij.lang.annotations.Language
@@ -15,13 +22,13 @@ import org.jooq.impl.DSL
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class DatabricksViewGeneratorTest {
+class SynapseViewGeneratorTest {
 
-    private lateinit var underTest: DatabricksViewGenerator
+    private lateinit var underTest: SynapseViewGenerator
 
     @BeforeEach
     fun setUp() {
-        underTest = DatabricksViewGenerator(DataPolicy.getDefaultInstance())
+        underTest = SynapseViewGenerator(DataPolicy.getDefaultInstance())
     }
 
     @Test
@@ -33,7 +40,7 @@ class DatabricksViewGeneratorTest {
         val condition = underTest.toPrincipalCondition(principals)
 
         // Then
-        condition!!.toSql() shouldBe "((is_account_group_member('analytics')) or (is_account_group_member('marketing')))"
+        condition!!.toSql() shouldBe "((IS_ROLEMEMBER('analytics')=1) or (IS_ROLEMEMBER('marketing')=1))"
     }
 
     @Test
@@ -45,13 +52,13 @@ class DatabricksViewGeneratorTest {
         val condition = underTest.toPrincipalCondition(principals)
 
         // Then
-        condition!!.toSql() shouldBe "(is_account_group_member('analytics'))"
+        condition!!.toSql() shouldBe "(IS_ROLEMEMBER('analytics')=1)"
     }
 
     @Test
     fun `principal check without any principals`() {
         // Given
-        val principals = emptyList<DataPolicy.Principal>()
+        val principals = emptyList<Principal>()
 
         // When
         val condition = underTest.toPrincipalCondition(principals)
@@ -64,18 +71,18 @@ class DatabricksViewGeneratorTest {
     fun `field transform with three transforms`() {
         // Given
         val attribute = namedField("email", "string")
-        val fixed = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setFixed(DataPolicy.RuleSet.FieldTransform.Transform.Fixed.newBuilder().setValue("****"))
+        val fixed = FieldTransform.Transform.newBuilder()
+            .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("****"))
             .addAllPrincipals(listOf("marketing", "analytics").toPrincipals())
             .build()
-        val otherFixed = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setFixed(DataPolicy.RuleSet.FieldTransform.Transform.Fixed.newBuilder().setValue("REDACTED EMAIL"))
-            .addAllPrincipals(listOf("fraud-and-risk").toPrincipals())
+        val otherFixed = FieldTransform.Transform.newBuilder()
+            .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("REDACTED EMAIL"))
+            .addAllPrincipals(listOf("fraud_and_risk").toPrincipals())
             .build()
-        val fallbackTransform = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setFixed(DataPolicy.RuleSet.FieldTransform.Transform.Fixed.newBuilder().setValue("stoelpoot"))
+        val fallbackTransform = FieldTransform.Transform.newBuilder()
+            .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("stoelpoot"))
             .build()
-        val fieldTransform = DataPolicy.RuleSet.FieldTransform.newBuilder()
+        val fieldTransform = FieldTransform.newBuilder()
             .setField(attribute)
             .addAllTransforms(listOf(fixed, otherFixed, fallbackTransform))
             .build()
@@ -84,7 +91,7 @@ class DatabricksViewGeneratorTest {
         val field = underTest.toJooqField(attribute, fieldTransform)
 
         // Then
-        field.toSql() shouldBe "case when ((is_account_group_member('marketing')) or (is_account_group_member('analytics'))) then '****' when (is_account_group_member('fraud-and-risk')) then " +
+        field.toSql() shouldBe "case when ((IS_ROLEMEMBER('marketing')=1) or (IS_ROLEMEMBER('analytics')=1)) then '****' when (IS_ROLEMEMBER('fraud_and_risk')=1) then " +
                 "'REDACTED EMAIL' else 'stoelpoot' end \"email\""
     }
 
@@ -92,14 +99,14 @@ class DatabricksViewGeneratorTest {
     fun `field transform with two transforms`() {
         // Given
         val attribute = namedField("email", "string")
-        val fixed = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setFixed(DataPolicy.RuleSet.FieldTransform.Transform.Fixed.newBuilder().setValue("****"))
+        val fixed = FieldTransform.Transform.newBuilder()
+            .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("****"))
             .addPrincipals("analytics".toPrincipal())
             .build()
-        val fallbackTransform = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setFixed(DataPolicy.RuleSet.FieldTransform.Transform.Fixed.newBuilder().setValue("stoelpoot"))
+        val fallbackTransform = FieldTransform.Transform.newBuilder()
+            .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("stoelpoot"))
             .build()
-        val fieldTransform = DataPolicy.RuleSet.FieldTransform.newBuilder()
+        val fieldTransform = FieldTransform.newBuilder()
             .setField(attribute)
             .addAllTransforms(listOf(fixed, fallbackTransform))
             .build()
@@ -108,20 +115,19 @@ class DatabricksViewGeneratorTest {
         val field = underTest.toJooqField(attribute, fieldTransform)
 
         // Then
-        field.toSql() shouldBe "case when (is_account_group_member('analytics')) then '****' else 'stoelpoot' end \"email\""
+        field.toSql() shouldBe "case when (IS_ROLEMEMBER('analytics')=1) then '****' else 'stoelpoot' end \"email\""
     }
 
     @Test
     fun `field transform with single transform`() {
         // Given
         val attribute = namedField("email")
-        val fallbackTransform = DataPolicy.RuleSet.FieldTransform.Transform.newBuilder()
-            .setRegexp(
-                DataPolicy.RuleSet.FieldTransform.Transform.Regexp.newBuilder().setRegexp("^.*(@.*)$")
-                    .setReplacement("****$1")
+        val fallbackTransform = FieldTransform.Transform.newBuilder()
+            .setFixed(
+                FieldTransform.Transform.Fixed.newBuilder().setValue("former regexp")
             )
             .build()
-        val fieldTransform = DataPolicy.RuleSet.FieldTransform.newBuilder()
+        val fieldTransform = FieldTransform.newBuilder()
             .setField(attribute)
             .addTransforms(fallbackTransform)
             .build()
@@ -130,7 +136,7 @@ class DatabricksViewGeneratorTest {
         val field = underTest.toJooqField(attribute, fieldTransform)
 
         // Then
-        field.toSql() shouldBe "regexp_replace(email, '^.*(@.*)\$', '****\$1') \"email\""
+        field.toSql() shouldBe "'former regexp' \"email\""
     }
 
     @Test
@@ -154,7 +160,7 @@ class DatabricksViewGeneratorTest {
                     .addAllConditions(
                         listOf(
                             GenericFilter.Condition.newBuilder()
-                                .addAllPrincipals(listOf("fraud-and-risk").toPrincipals())
+                                .addAllPrincipals(listOf("fraud_and_risk").toPrincipals())
                                 .setCondition("true")
                                 .build(),
                             GenericFilter.Condition.newBuilder()
@@ -173,27 +179,35 @@ class DatabricksViewGeneratorTest {
         val condition = underTest.toCondition(filter.genericFilter)
 
         // Then
-        condition.toSql() shouldBe "case when (is_account_group_member('fraud-and-risk')) then true when ((is_account_group_member('analytics')) or (is_account_group_member('marketing'))) " +
-                "then age > 18 else false end"
+        condition.toSql() shouldBe """(1 = (case when (IS_ROLEMEMBER('fraud_and_risk')=1) then (case
+  when (1=1) then 1
+  else 0
+end) when ((IS_ROLEMEMBER('analytics')=1) or (IS_ROLEMEMBER('marketing')=1)) then (case
+  when (age > 18) then 1
+  else 0
+end) else (case
+  when (1=0) then 1
+  else 0
+end) end))""".trimMargin()
     }
 
     @Test
     fun `single retention to condition`() {
         // Given
-        val retention = DataPolicy.RuleSet.Filter.RetentionFilter.newBuilder()
+        val retention = RetentionFilter.newBuilder()
             .setField(DataPolicy.Field.newBuilder().addNameParts("timestamp"))
             .addAllConditions(
                 listOf(
-                    DataPolicy.RuleSet.Filter.RetentionFilter.Condition.newBuilder()
-                        .addPrincipals(DataPolicy.Principal.newBuilder().setGroup("marketing"))
-                        .setPeriod(DataPolicy.RuleSet.Filter.RetentionFilter.Period.newBuilder().setDays(5))
+                    RetentionFilter.Condition.newBuilder()
+                        .addPrincipals(Principal.newBuilder().setGroup("marketing"))
+                        .setPeriod(RetentionFilter.Period.newBuilder().setDays(5))
                         .build(),
-                    DataPolicy.RuleSet.Filter.RetentionFilter.Condition.newBuilder()
-                        .addPrincipals(DataPolicy.Principal.newBuilder().setGroup("fraud-and-risk"))
+                    RetentionFilter.Condition.newBuilder()
+                        .addPrincipals(Principal.newBuilder().setGroup("fraud_and_risk"))
                         .build(),
-                    DataPolicy.RuleSet.Filter.RetentionFilter.Condition.newBuilder()
-                        .addPrincipals(DataPolicy.Principal.getDefaultInstance())
-                        .setPeriod(DataPolicy.RuleSet.Filter.RetentionFilter.Period.newBuilder().setDays(10))
+                    RetentionFilter.Condition.newBuilder()
+                        .addPrincipals(Principal.getDefaultInstance())
+                        .setPeriod(RetentionFilter.Period.newBuilder().setDays(10))
                         .build()
                 )
             ).build()
@@ -203,17 +217,21 @@ class DatabricksViewGeneratorTest {
 
         // Then
         condition.toSql() shouldBe """
-            case when (is_account_group_member('marketing')) then dateadd(day, 5, timestamp) > current_timestamp when (is_account_group_member('fraud-and-risk')) then true else dateadd(day, 10, timestamp) > current_timestamp end""".trimIndent()
+            dateadd(day, (case
+              when (IS_ROLEMEMBER('marketing')=1) then 5
+              when (IS_ROLEMEMBER('fraud_and_risk')=1) then 100000
+              else 10
+            end), timestamp) > current_timestamp""".trimIndent()
     }
 
     @Test
     fun `full sql view statement with multiple retentions`() {
         // Given
-        val viewGenerator = DatabricksViewGenerator(multipleRetentionPolicy) { withRenderFormatted(true) }
+        val viewGenerator = SynapseViewGenerator(multipleRetentionPolicy) { withRenderFormatted(true) }
         // When
 
         // Then
-        viewGenerator.toDynamicViewSQL().sql shouldBe """create or replace view public.demo_view
+        viewGenerator.toDynamicViewSQL().sql shouldBe """create view public.demo_view
 as
 select
   ts,
@@ -222,30 +240,36 @@ select
   transactionamount
 from public.demo_tokenized
 where (
-  case
-    when (is_account_group_member('fraud-and-risk')) then true
-    else transactionamount < 10
-  end
-  and case
-    when (is_account_group_member('marketing')) then dateadd(day, 5, ts) > current_timestamp
-    when (is_account_group_member('fraud-and-risk')) then true
-    else dateadd(day, 10, ts) > current_timestamp
-  end
-  and case
-    when (is_account_group_member('fraud-and-risk')) then dateadd(day, 365, validThrough) > current_timestamp
-    else dateadd(day, 0, validThrough) > current_timestamp
-  end
+  (1 = (case
+    when (IS_ROLEMEMBER('fraud_and_risk')=1) then (case
+  when (1=1) then 1
+  else 0
+end)
+    else (case
+  when (transactionamount < 10) then 1
+  else 0
+end)
+  end))
+  and dateadd(day, (case
+  when (IS_ROLEMEMBER('marketing')=1) then 5
+  when (IS_ROLEMEMBER('fraud_and_risk')=1) then 100000
+  else 10
+end), ts) > current_timestamp
+  and dateadd(day, (case
+  when (IS_ROLEMEMBER('fraud_and_risk')=1) then 365
+  else 0
+end), validThrough) > current_timestamp
 );"""
     }
 
     @Test
     fun `full sql view statement with single retention`() {
         // Given
-        val viewGenerator = DatabricksViewGenerator(singleRetentionPolicy) { withRenderFormatted(true) }
+        val viewGenerator = SynapseViewGenerator(singleRetentionPolicy) { withRenderFormatted(true) }
         // When
 
         // Then
-        viewGenerator.toDynamicViewSQL().sql shouldBe """create or replace view public.demo_view
+        viewGenerator.toDynamicViewSQL().sql shouldBe """create view public.demo_view
 as
 select
   ts,
@@ -253,39 +277,45 @@ select
   transactionamount
 from public.demo_tokenized
 where (
-  case
-    when (is_account_group_member('fraud-and-risk')) then true
-    else transactionamount < 10
-  end
-  and case
-    when (is_account_group_member('marketing')) then dateadd(day, 5, ts) > current_timestamp
-    when (is_account_group_member('fraud-and-risk')) then true
-    else dateadd(day, 10, ts) > current_timestamp
-  end
+  (1 = (case
+    when (IS_ROLEMEMBER('fraud_and_risk')=1) then (case
+  when (1=1) then 1
+  else 0
+end)
+    else (case
+  when (transactionamount < 10) then 1
+  else 0
+end)
+  end))
+  and dateadd(day, (case
+  when (IS_ROLEMEMBER('marketing')=1) then 5
+  when (IS_ROLEMEMBER('fraud_and_risk')=1) then 100000
+  else 10
+end), ts) > current_timestamp
 );"""
     }
 
     @Test
     fun `transform test various transforms`() {
-        underTest = DatabricksViewGenerator(dataPolicy) { withRenderFormatted(true) }
+        underTest = SynapseViewGenerator(policyWithoutRegexes) { withRenderFormatted(true) }
         underTest.toDynamicViewSQL().sql
             .shouldBe(
-                """create or replace view my_catalog.my_schema.gddemo_public
+                """create view my_schema.gddemo_public
 as
 select
   transactionId,
   case
-    when (is_account_group_member('fraud-and-risk')) then userId
-    else hash(1234, userId)
+    when (IS_ROLEMEMBER('fraud_and_risk')=1) then userId
+    else HASHBYTES('SHA2_512', userId)
   end userId,
   case
     when (
-      (is_account_group_member('analytics'))
-      or (is_account_group_member('marketing'))
-    ) then regexp_replace(email, '^.*(@.*)${'$'}', '****${'$'}1')
+      (IS_ROLEMEMBER('analytics')=1)
+      or (IS_ROLEMEMBER('marketing')=1)
+    ) then 'former regexp'
     when (
-      (is_account_group_member('fraud-and-risk'))
-      or (is_account_group_member('admin'))
+      (IS_ROLEMEMBER('fraud_and_risk')=1)
+      or (IS_ROLEMEMBER('admin')=1)
     ) then email
     else '****'
   end email,
@@ -297,43 +327,76 @@ select
   itemCount,
   date,
   purpose
-from mycatalog.my_schema.gddemo
+from my_schema.gddemo
 where (
-  case
-    when (is_account_group_member('fraud-and-risk')) then true
-    else age > 18
-  end
-  and case
-    when (is_account_group_member('marketing')) then userId in ('1', '2', '3', '4')
-    else true
-  end
-  and transactionAmount < 10
+  (1 = (case
+    when (IS_ROLEMEMBER('fraud_and_risk')=1) then (case
+  when (1=1) then 1
+  else 0
+end)
+    else (case
+  when (age > 18) then 1
+  else 0
+end)
+  end))
+  and (1 = (case
+    when (IS_ROLEMEMBER('marketing')=1) then (case
+  when (userId in ('1', '2', '3', '4')) then 1
+  else 0
+end)
+    else (case
+  when (1=1) then 1
+  else 0
+end)
+  end))
+  and (1 = ((case
+  when (transactionAmount < 10) then 1
+  else 0
+end)))
 );"""
+            )
+    }
+
+    @Test
+    fun `grant select permission`() {
+        underTest = SynapseViewGenerator(policyWithoutRegexes) { withRenderFormatted(true) }
+        underTest.grantSelectPrivileges().sql
+            .shouldBe(
+                """grant SELECT on my_schema.gddemo_public to "analytics";
+grant SELECT on my_schema.gddemo_public to "marketing";
+grant SELECT on my_schema.gddemo_public to "fraud_and_risk";
+grant SELECT on my_schema.gddemo_public to "admin";"""
             )
     }
 
     @Test
     fun `transform - no row filters`() {
         val policyWithoutFilters = dataPolicy.toBuilder().apply { ruleSetsBuilderList.first().clearFilters() }.build()
-        underTest = DatabricksViewGenerator(policyWithoutFilters) { withRenderFormatted(true) }
+        underTest = SynapseViewGenerator(policyWithoutFilters) { withRenderFormatted(true) }
+        shouldThrow<BadRequestException> { underTest.toDynamicViewSQL() }
+    }
+
+    @Test
+    fun `transform without regex replace - no row filters`() {
+        underTest = SynapseViewGenerator(policyWithoutFiltersAndRegexes) { withRenderFormatted(true) }
         underTest.toDynamicViewSQL().sql
             .shouldBe(
-                """create or replace view my_catalog.my_schema.gddemo_public
+                """create view my_schema.gddemo_public
 as
 select
   transactionId,
   case
-    when (is_account_group_member('fraud-and-risk')) then userId
-    else hash(1234, userId)
+    when (IS_ROLEMEMBER('fraud_and_risk')=1) then userId
+    else HASHBYTES('SHA2_512', userId)
   end userId,
   case
     when (
-      (is_account_group_member('analytics'))
-      or (is_account_group_member('marketing'))
-    ) then regexp_replace(email, '^.*(@.*)${'$'}', '****${'$'}1')
+      (IS_ROLEMEMBER('analytics')=1)
+      or (IS_ROLEMEMBER('marketing')=1)
+    ) then 'former regexp'
     when (
-      (is_account_group_member('fraud-and-risk'))
-      or (is_account_group_member('admin'))
+      (IS_ROLEMEMBER('fraud_and_risk')=1)
+      or (IS_ROLEMEMBER('admin')=1)
     ) then email
     else '****'
   end email,
@@ -345,8 +408,55 @@ select
   itemCount,
   date,
   purpose
-from mycatalog.my_schema.gddemo;"""
+from my_schema.gddemo;"""
             )
+    }
+
+
+    @Test
+    fun `single generic condition`() {
+        val f = GenericFilter.newBuilder()
+            .addAllConditions(
+                listOf(
+                    GenericFilter.Condition.newBuilder()
+                        .addAllPrincipals(groupPrincipals("fraud_and_risk"))
+                        .setCondition("1=1")
+                        .build(),
+                    GenericFilter.Condition.newBuilder()
+                        .addAllPrincipals(groupPrincipals("marketing", "sales"))
+                        .setCondition("transactionamount < 100")
+                        .build(),
+                    GenericFilter.Condition.newBuilder()
+                        .setCondition("transactionamount < 10")
+                        .build()
+                )
+            )
+            .build()
+        val sql = underTest.toCondition(f)
+        println(sql.toSql())
+    }
+
+    @Test
+    fun `single retention condition`() {
+        val f = RetentionFilter.newBuilder()
+            .setField(DataPolicy.Field.newBuilder().addNameParts("ts"))
+            .addAllConditions(
+                listOf(
+                    RetentionFilter.Condition.newBuilder()
+                        .addAllPrincipals(groupPrincipals("fraud_and_risk"))
+                        .build(),
+                    RetentionFilter.Condition.newBuilder()
+                        .addAllPrincipals(groupPrincipals("marketing", "sales"))
+                        .setPeriod(RetentionFilter.Period.newBuilder().setDays(3))
+                        .build(),
+                    RetentionFilter.Condition.newBuilder()
+                        .setPeriod(RetentionFilter.Period.newBuilder().setDays(10))
+                        .build(),
+                )
+            )
+            .build()
+        val sql = underTest.toCondition(f)
+        println(sql.toSql())
     }
 
     companion object {
@@ -409,7 +519,7 @@ from mycatalog.my_schema.gddemo;"""
                 regexp: '^.*(@.*)${'$'}'
                 replacement: '****${'$'}1'
             - principals:
-                - group: fraud-and-risk
+                - group: fraud_and_risk
                 - group: admin
               identity: {}
             - principals: []
@@ -420,7 +530,7 @@ from mycatalog.my_schema.gddemo;"""
               - userId
           transforms:
             - principals:
-                - group: fraud-and-risk
+                - group: fraud_and_risk
               identity: {}
             - principals: []
               hash:
@@ -440,19 +550,13 @@ from mycatalog.my_schema.gddemo;"""
                 statement: "case when brand = 'Macbook' then 'Apple' else 'Other' end"
       filters:
         - generic_filter:
-            field:
-              name_parts:
-                - age
             conditions:
               - principals:
-                  - group: fraud-and-risk
+                  - group: fraud_and_risk
                 condition: "true"
               - principals: []
                 condition: "age > 18"
         - generic_filter:
-            field:
-              name_parts:
-                - userId
             conditions:
               - principals:
                   - group: marketing
@@ -460,19 +564,41 @@ from mycatalog.my_schema.gddemo;"""
               - principals: []
                 condition: "true"
         - generic_filter:
-            field:
-              name_parts:
-                - transactionAmount
             conditions:
               - principals: []
                 condition: "transactionAmount < 10"
     info:
-      title: "Data Policy for Pace Databricks Demo Dataset"
+      title: "Data Policy for Pace Synapse Demo Dataset"
       description: "Pace Demo Dataset"
       version: "1.0.0"
       create_time: "2023-09-26T16:33:51.150Z"
       update_time: "2023-09-26T16:33:51.150Z"
               """.yaml2json().parseDataPolicy()
+
+        val policyWithoutRegexes: DataPolicy = dataPolicy.toBuilder().apply {
+            val fieldTransforms = this.ruleSetsBuilderList.first().fieldTransformsList.map { fieldTransform ->
+                FieldTransform.newBuilder()
+                    .setField(fieldTransform.field)
+                    .addAllTransforms(
+                        fieldTransform.transformsList.map {
+                            if (it.hasRegexp()) {
+                                it.toBuilder().clearRegexp()
+                                    .setFixed(FieldTransform.Transform.Fixed.newBuilder().setValue("former regexp"))
+                                    .build()
+                            } else {
+                                it
+                            }
+                        }
+                    ).build()
+            }
+            ruleSetsBuilderList.first().clearFieldTransforms()
+                .addAllFieldTransforms(fieldTransforms)
+        }.build()
+
+        val policyWithoutFiltersAndRegexes: DataPolicy = policyWithoutRegexes.toBuilder().apply {
+            this.ruleSetsBuilderList.first().clearFilters()
+        }.build()
+
 
         @Language("yaml")
         val singleRetentionPolicy = """
@@ -504,7 +630,7 @@ from mycatalog.my_schema.gddemo;"""
                 filters:
                   - generic_filter:
                       conditions:
-                        - principals: [ {group: fraud-and-risk} ]
+                        - principals: [ {group: fraud_and_risk} ]
                           condition: "true"
                         - principals : []
                           condition: "transactionamount < 10"
@@ -518,7 +644,7 @@ from mycatalog.my_schema.gddemo;"""
                         - principals: [ {group: marketing} ]
                           period:
                             days: 5
-                        - principals: [ {group: fraud-and-risk} ]
+                        - principals: [ {group: fraud_and_risk} ]
                         - principals: [] 
                           period:
                             days: 10
@@ -558,7 +684,7 @@ from mycatalog.my_schema.gddemo;"""
                 filters:
                   - generic_filter:
                       conditions:
-                        - principals: [ {group: fraud-and-risk} ]
+                        - principals: [ {group: fraud_and_risk} ]
                           condition: "true"
                         - principals : []
                           condition: "transactionamount < 10"
@@ -572,7 +698,7 @@ from mycatalog.my_schema.gddemo;"""
                         - principals: [ {group: marketing} ]
                           period:
                             days: 5
-                        - principals: [ {group: fraud-and-risk} ]
+                        - principals: [ {group: fraud_and_risk} ]
                         - principals: [] 
                           period:
                             days: 10
@@ -583,7 +709,7 @@ from mycatalog.my_schema.gddemo;"""
                         required: true
                         type: timestamp
                       conditions:
-                        - principals: [ {group: fraud-and-risk} ]
+                        - principals: [ {group: fraud_and_risk} ]
                           period:
                             days: 365
                         - principals: [] 
@@ -591,4 +717,12 @@ from mycatalog.my_schema.gddemo;"""
                             days: 0
         """.trimIndent().yaml2json().parseDataPolicy()
     }
+
+
+    private fun groupPrincipals(vararg p: String): List<Principal> =
+        p.map {
+            Principal.newBuilder()
+                .setGroup(it)
+                .build()
+        }
 }